requirements.txt
# Web & Queue
Flask>=3.0.2
celery>=5.3.6
redis>=5.0.3
pymongo>=4.6.2
python-dotenv>=1.0.1
#gunicorn>=21.2.0

# Machine Learning & Media Processing
torch>=2.6.0
transformers>=4.40.0
sentence-transformers>=3.0.0
faster-whisper>=1.0.1
peft>=0.10.0
librosa>=0.10.1
numpy>=1.26.4
pandas>=2.2.1

# System Utils
psutil>=5.9.8

#media ingestion
yt-dlp>=2025.01.01